backend/src/requirements/base.txt
Django==6.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1